# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'QASA API Documentation'
  description: 'API documentation for QASA Laravel application - invoicing and client management system'
  contact:
    name: 'QASA Support'
    email: support@qasa.sk
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
paths:
...
  /api/v1/quotes:
    get:
      tags:
        - Quotes
      summary: 'List quotes'
      operationId: 731541b0d8b53d7b5331bd4da94f0726
      responses:
        200:
          description: 'List of quotes'
      security:
        -
          sanctum: []
    post:
      tags:
        - Quotes
      summary: 'Create quote'
      operationId: bd42a4022cc87432544fc194dab26301
      responses:
        201:
          description: 'Quote created'
        422:
          description: 'Validation error'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{id}':
    get:
      tags:
        - Quotes
      summary: 'Get quote details'
      operationId: 51c9556c69991ade22f1013045c97e58
      responses:
        200:
          description: 'Quote details'
        404:
          description: 'Quote not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Quotes
      summary: 'Update draft quote header'
      operationId: f23e54fa00cf117c70146d6a9a9beeb8
      responses:
        200:
          description: 'Quote updated'
        422:
          description: 'Validation error or quote not editable'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Quotes
      summary: 'Delete draft quote'
      operationId: 9863550550d4ac22f6b70e9be265a5c0
      responses:
        204:
          description: 'Quote deleted'
        403:
          description: 'Only draft quotes can be deleted'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{quote}/items':
    post:
      tags:
        - Quotes
      summary: 'Add item to quote'
      operationId: 92bfcec86b3f14c9e0527910b80e115f
      responses:
        201:
          description: 'Item added'
        422:
          description: 'Validation error or quote not editable'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{quote}/items/{item}':
    delete:
      tags:
        - Quotes
      summary: 'Remove item from quote'
      operationId: 202078c0fe41efcfe85b952d762a2912
      responses:
        204:
          description: 'Item removed'
        422:
          description: 'Quote not editable'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{quote}/status':
    post:
      tags:
        - Quotes
      summary: 'Manually update quote status'
      operationId: 816351600c90ae5b21e84fd53ac05243
      responses:
        200:
          description: 'Quote status updated'
        422:
          description: 'Invalid status transition'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{quote}/email':
    post:
      tags:
        - Quotes
      summary: 'Email the quote PDF to the client (sends the quote first when still a draft)'
      operationId: 1b5f1bb6320cae49ea3e222c2d555fe2
      responses:
        200:
          description: 'Quote queued for email delivery'
        422:
          description: 'Validation error or client without email'
        429:
          description: 'Too many email requests'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{quote}/pdf/download':
    get:
      tags:
        - Quotes
      summary: 'Download quote PDF'
      operationId: fe001c5716d11d3728367cd60d08e3ba
      responses:
        200:
          description: 'PDF file download'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{quote}/pdf/preview':
    get:
      tags:
        - Quotes
      summary: 'Preview quote PDF in browser'
      operationId: 36a4a8f9b944524f5913ce480a24696a
      responses:
        200:
          description: 'PDF preview'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{quote}/public-link':
    post:
      tags:
        - Quotes
      summary: 'Create (or return the existing) public link for the quote; pass regenerate=true for a fresh token'
      operationId: 659fe617520fe069abf28d30777fd967
      responses:
        200:
          description: 'Public link token and URL'
        422:
          description: 'Quote is still a draft'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Quotes
      summary: 'Revoke the quote public link'
      operationId: 7b869bed02f35bc7d4c830cad9eac765
      responses:
        204:
          description: 'Public link revoked'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{quote}/convert-to-invoice':
    post:
      tags:
        - Quotes
      summary: 'Convert an accepted (or sent) quote into a draft invoice'
      operationId: 4ccd28b747e246604fec64dc7cd00d1d
      responses:
        201:
          description: 'Invoice created'
        422:
          description: 'Invalid status or already converted'
      security:
        -
          sanctum: []
  '/api/v1/quotes/{quote}/convert-to-order':
    post:
      tags:
        - Quotes
      summary: 'Convert an accepted (or sent) quote into an order'
      operationId: 48216424d363dd9307acd0aa8e646147
      responses:
        201:
          description: 'Order created'
        422:
          description: 'Invalid status or already converted'
      security:
        -
          sanctum: []
...
components:
...
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header
...
security:
  -
    sanctum: []